[KMP] Move ArPlacement enum to pdp domain
